+++ Assets/Hertzole/Gold Player/Scripts/Runtime/Core/Classes/InputItem.cs
@@ -1,4 +1,4 @@
-﻿#if ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT // Mark this as obsolete if the new input system is enabled.
+﻿#if ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT && !ENABLE_LEGACY_INPUT_MANAGER // Mark this as obsolete if the new input system is enabled.
 #define OBSOLETE
 #endif
 
@@ -23,14 +23,22 @@ namespace Hertzole.GoldPlayer
     [Serializable]
     public struct InputItem : IEquatable<InputItem>
     {
+        public enum InputType { Button = 0, Axis = 1, Vector2 = 2 }
+
         [SerializeField]
         [Tooltip("The name code will reference the item with.")]
         [FormerlySerializedAs("m_ButtonName")]
         private string buttonName;
         [SerializeField]
+        [Tooltip("The type of input this will be.")]
+        private InputType type;
+        [SerializeField]
         [Tooltip("The name in the Input Manager.")]
         [FormerlySerializedAs("m_InputName")]
         private string inputName;
+        [SerializeField]
+        [Tooltip("The second name in the Input Manager. This is only used for Vector2 values.")]
+        private string inputNameSecondary;
         [SerializeField]
         [Tooltip("The key code for the item.")]
         [FormerlySerializedAs("m_Key")]
@@ -38,8 +46,12 @@ namespace Hertzole.GoldPlayer
 
         /// <summary> The name code will reference the item with. </summary>
         public string ButtonName { get { return buttonName; } set { buttonName = value; } }
+        /// <summary> The type of input this will be. </summary>
+        public InputType Type { get { return type; } set { type = value; } }
         /// <summary> The name in the Input Manager. </summary>
         public string InputName { get { return inputName; } set { inputName = value; } }
+        /// <summary> The second name in the Input Manager. This is only used for Vector2 values. </summary>
+        public string InputNameSecondary { get { return inputNameSecondary; } set { inputNameSecondary = value; } }
         /// <summary> The key code for the item. </summary>
         public KeyCode Key { get { return key; } set { key = value; } }
 
@@ -47,6 +59,27 @@ namespace Hertzole.GoldPlayer
         {
             this.buttonName = buttonName;
             this.inputName = inputName;
+            this.key = key;
+
+            type = InputType.Button;
+            inputNameSecondary = null;
+        }
+
+        public InputItem(string buttonName, string xAxis, string yAxis)
+        {
+            this.buttonName = buttonName;
+            inputName = xAxis;
+            inputNameSecondary = yAxis;
+            type = InputType.Vector2;
+            key = KeyCode.None;
+        }
+
+        public InputItem(string buttonName, InputType type, string inputName, string inputNameSecondary, KeyCode key)
+        {
+            this.buttonName = buttonName;
+            this.type = type;
+            this.inputName = inputName;
+            this.inputNameSecondary = inputNameSecondary;
             this.key = key;
         }