+++ Assets/Hertzole/Gold Player/Scripts/Runtime/Core/Modules/PlayerCamera.cs
@@ -59,23 +59,8 @@ namespace Hertzole.GoldPlayer
         private Transform cameraHead = null;
 
         [SerializeField]
-#if !ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT
-        [HideInInspector]
-#endif
         [Tooltip("Look action for the new Input System.")]
         private string input_Look = "Look";
-        [SerializeField]
-#if ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT
-        [HideInInspector]
-#endif
-        [Tooltip("Mouse X axis for the old Input Manager.")]
-        private string input_MouseX = "Mouse X";
-        [SerializeField]
-#if ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT
-        [HideInInspector]
-#endif
-        [Tooltip("Mouse Y axis for the old Input Manager.")]
-        private string input_MouseY = "Mouse Y";
 
         // Determines if a camera shake should be preformed.
         private bool doShake = false;
@@ -169,10 +154,6 @@ namespace Hertzole.GoldPlayer
 
         /// <summary> Look action for the new Input System. </summary>
         public string LookInput { get { return input_Look; } set { input_Look = value; } }
-        /// <summary> Mouse X axis for the old Input Manager. </summary>
-        public string MouseX { get { return input_MouseX; } set { input_MouseX = value; } }
-        /// <summary> Mouse Y axis for the old Input Manager. </summary>
-        public string MouseY { get { return input_MouseY; } set { input_MouseY = value; } }
 
         /// <summary> Where the head should be looking. </summary>
         public Vector3 TargetHeadAngles { get { return targetHeadAngles; } }
@@ -204,6 +185,12 @@ namespace Hertzole.GoldPlayer
 #if UNITY_EDITOR
         [System.Obsolete("Use 'FieldOfViewKick' instead. This will be removed on build.", true)]
         public FOVKickClass FOVKick { get { return fieldOfViewKick; } set { fieldOfViewKick = value; } }
+        /// <summary> Mouse X axis for the old Input Manager. </summary>
+        [System.Obsolete("Use 'LookInput' instead along with GetVector2. This will be removed on build.", true)]
+        public string MouseX { get { return null; } set { } }
+        /// <summary> Mouse Y axis for the old Input Manager. </summary>
+        [System.Obsolete("Use 'LookInput' instead along with GetVector2. This will be removed on build.", true)]
+        public string MouseY { get { return null; } set { } }
 #endif
         #endregion
 
@@ -278,7 +265,6 @@ namespace Hertzole.GoldPlayer
             if (canLookAround && !forceLooking)
             {
                 // Set the input.
-#if ENABLE_INPUT_SYSTEM && GOLD_PLAYER_NEW_INPUT
                 mouseInput = GetVector2Input(input_Look);
                 if (invertXAxis)
                 {
@@ -289,9 +275,6 @@ namespace Hertzole.GoldPlayer
                 {
                     mouseInput.y = -mouseInput.y;
                 }
-#else
-                mouseInput = new Vector2(invertXAxis ? -GetAxis(input_MouseX) : GetAxis(input_MouseX), invertYAxis ? -GetAxis(input_MouseY) : GetAxis(input_MouseY));
-#endif
             }
             else
             {